requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
Flask-SocketIO>=5.3,<6
simple-websocket>=1.0,<2
google-genai>=1.0.0,<2
psycopg[binary]>=3.1,<4
gunicorn>=22,<24
requests>=2.31,<3
